# Conflicts:
#	inc/server/server.chars.php
#	inc/server/server.commands.php
#	install/sql/full_install.sql

+++ core/class.database.php
@@ -140,7 +140,7 @@ class Database
     {
         $sql = mysqli_query($this->mysql, $query) or die("Couldnt Run Query: ".$query."<br />Error: ".mysqli_error($this->mysql)."");
 		$this->_statistics['count']++;
-		return mysqli_fetch_assoc($sql);
+		return (int)mysqli_fetch_assoc($sql);
     }
 
 //	************************************************************
